Test master

+++ src/Template.Web/Areas/Example/Timesheets/TimesheetsController.cs
@@ -57,88 +57,81 @@ namespace Template.Web.Areas.Example.Timesheets
             return View(model);
         }
 
-
-[HttpPost]
-        [ValidateAntiForgeryToken]
-        public virtual async Task<IActionResult> Edit2(EditViewModel model, string WeekDaysEncoded)
+        [HttpGet]
+        public virtual async Task<IActionResult> GetUsers(Guid timesheetId)
         {
-            Console.WriteLine($"POST Edit ricevuto - Name: {model.Name}, WeekDaysEncoded: {WeekDaysEncoded}");
-            
-            if (!string.IsNullOrEmpty(WeekDaysEncoded))
+            var teamMembersDto = await _sharedService.Query(new TimesheetsIndexQuery
             {
-                model.WeekDay = WeekDaysEncoded;
-            }
-
-            if (!ModelState.IsValid)
+                IdCurrentTimesheet = timesheetId,
+                Paging = null
+            });
+            var usersDto = await _sharedService.Query(new UsersIndexQuery
             {
-                Console.WriteLine("ModelState non valido:");
-                foreach(var e in ModelState.Values.SelectMany(v => v.Errors))
+                IdCurrentUser = Guid.Empty,
+                Paging = null
+            });
+            var userDictionary = usersDto.Users
+                .Where(tm => tm.TimesheetId == timesheetId)
+                .Select(tm =>
                 {
-                    Console.WriteLine(e.ErrorMessage);
-                }
-                Alerts.AddError(this, "Errore in aggiornamento");
-                return RedirectToAction(nameof(Index));
-            }
+                    return new
+                    {
+                        userId = tm.Id,
+                        firstName = tm.FirstName,
+                        lastName = tm.LastName,
+                        email = tm.Email,
+                        role = tm.Role
+                    };
+                })
+                .ToList();
 
-            try
-            {
-                Console.WriteLine("Prima di chiamare HandleTimesheet");
-                model.Id = await _sharedService.HandleTimesheet(model.ToAddOrUpdateTimesheetCommand());
-                Console.WriteLine("Dopo chiamata HandleTimesheet");
-                Alerts.AddSuccess(this, "Informazioni aggiornate");
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Eccezione HandleTimesheet: " + e.Message);
-                ModelState.AddModelError(string.Empty, e.Message);
-                Alerts.AddError(this, "Errore in aggiornamento");
-            }
-
-            return RedirectToAction(nameof(Index));
+            return Json(userDictionary);
         }
 
-[HttpPost]
-[ValidateAntiForgeryToken]
-public virtual async Task<IActionResult> Edit(EditViewModel model, string WeekDaysEncoded)
-{
-    Console.WriteLine($"POST Edit ricevuto - Name: {model.Name}, WeekDay: {model.WeekDay}, StartTime: {model.StartTime}, EndTime: {model.EndTime}");
-
-    if (!string.IsNullOrEmpty(WeekDaysEncoded))
-    {
-        model.WeekDay = WeekDaysEncoded;
-
-    }
-
-    if (string.IsNullOrEmpty(model.WeekDay))
-    {
-        ModelState.AddModelError(nameof(model.WeekDay), "Seleziona almeno un giorno.");
-    }
-
-    if (!ModelState.IsValid)
-    {
-        Alerts.AddError(this, "Errore in aggiornamento");
-        // Se non vuoi creare una view Edit.cshtml, fai redirect a Index (oppure gestisci diversamente)
-        return RedirectToAction(nameof(Index));
-    }
-
-   try
-    {
-        Console.WriteLine("Prima di chiamare HandleTimesheet");
-        Console.WriteLine(model.StartTime);
-        Console.WriteLine(model.EndTime);
-        model.Id = await _sharedService.HandleTimesheet(model.ToAddOrUpdateTimesheetCommand());
-        Console.WriteLine("Dopo chiamata HandleTimesheet");
-        Alerts.AddSuccess(this, "Informazioni aggiornate");
-    }
-    catch (Exception e)
-    {
-        Console.WriteLine("Eccezione HandleTimesheet: " + e.Message);
-        ModelState.AddModelError(string.Empty, e.Message);
-        Alerts.AddError(this, "Errore in aggiornamento");
-    }
-    Console.WriteLine("Uscendo");
-    return RedirectToAction(nameof(Index));
-}
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public virtual async Task<IActionResult> Edit(EditViewModel model, string WeekDaysEncoded)
+        {   
+            var allTimesheet = await _sharedService.Query(new TimesheetsIndexQuery
+            {
+                IdCurrentTimesheet = Guid.Empty,
+                Paging = null
+            });
+            var timesheets = allTimesheet.Timesheets.Where(tm => tm.Name == model.Name && tm.Id != model.Id);
+            if (timesheets.Any() == true){
+                Alerts.AddError(this, "È già presente un timesheet con questo nome");
+            }
+            else{
+                Console.WriteLine($"POST Edit ricevuto - Name: {model.Name}, WeekDay: {model.WeekDay}, StartTime: {model.StartTime}, EndTime: {model.EndTime}");
+                if (!string.IsNullOrEmpty(WeekDaysEncoded))
+                {
+                    model.WeekDay = WeekDaysEncoded;
+                }
+                if (string.IsNullOrEmpty(model.WeekDay))
+                {
+                    ModelState.AddModelError(nameof(model.WeekDay), "Seleziona almeno un giorno.");
+                }
+                if (!ModelState.IsValid)
+                {
+                    Alerts.AddError(this, "Errore in aggiornamento");
+                    return RedirectToAction(nameof(Index));
+                }
+                
+                try
+                {
+                    model.Id = await _sharedService.HandleTimesheet(model.ToAddOrUpdateTimesheetCommand());
+                    Alerts.AddSuccess(this, "Informazioni aggiornate");
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine("Eccezione HandleTimesheet: " + e.Message);
+                    ModelState.AddModelError(string.Empty, e.Message);
+                    Alerts.AddError(this, "Errore in aggiornamento");
+                }
+                Console.WriteLine("Uscendo");
+            }
+            return RedirectToAction(nameof(Index));
+        }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -148,7 +141,18 @@ public virtual async Task<IActionResult> Edit(EditViewModel model, string WeekDa
 
             try
             {
-                await _sharedService.DeleteTimesheet(id); 
+                var usersWithTimesheet = await _sharedService.Query(new UsersByTimesheetQuery
+                {
+                    TimesheetId = id
+                });
+
+                if (usersWithTimesheet.Users.Any())
+                {
+                    Alerts.AddError(this, "Non puoi cancellare questo timesheet perché ci sono utenti associati.");
+                    return RedirectToAction(nameof(Index));
+                }
+                Console.WriteLine("ENTRATO IN DELETE - ID: " + id);
+                await _sharedService.DeleteTimesheet(id);
                 Alerts.AddSuccess(this, "Timesheet cancellato");
             }
             catch (Exception e)
